+++ lib/presentation/pages/admin_profile/financial_indicators_page.dart
@@ -145,7 +145,7 @@ class _LevelCards extends StatelessWidget {
       runSpacing: 16,
       children: List.generate(
         15,
-        (index) => LevelCard(
+        (index) => LevelCard.admin(
           level: index + 1,
           reward: index % 2 == 0 ? 300 : 500,
         ),
