+++ lib/presentation/pages/client_profile/sections/dashboard_tab.dart
@@ -104,7 +104,7 @@ class _LevelCards extends StatelessWidget {
       runSpacing: 16,
       children: List.generate(
         15,
-        (index) => LevelCard(
+        (index) => LevelCard.client(
           level: index + 1,
           reward: index % 2 == 0 ? 300 : 500,
         ),
